Move exception module
