+++ neo/exception.py
