+++ neo/parser.py
@@ -46,9 +46,9 @@ from typing_extensions import (
     get_type_hints,
 )
 
-from omc4py.exception import OMCRuntimeError
 from omc4py.string import unquote_modelica_string
 
+from .exception import OMCRuntimeError
 from .modelica import enumeration, record
 from .openmodelica import Component, TypeName, VariableName
